requirements.txt
streamlit==1.25.0
diffusers==0.31.0
transformers==4.46.2
huggingface-hub==0.26.2
torch==2.5.1
torchvision==0.20.1
Pillow==9.5.0
numpy==1.26.4